+++ previous-versions/banking/src/codatbanking/models/operations/get_transaction_category.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import transactioncategory as shared_transactioncategory
+from ...models.shared import transactioncategory as shared_transactioncategory
 from typing import Optional
 
 
@@ -23,10 +23,10 @@ class GetTransactionCategoryRequest:
 class GetTransactionCategoryResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     transaction_category: Optional[shared_transactioncategory.TransactionCategory] = dataclasses.field(default=None)
     r"""Success"""