+++ previous-versions/banking/src/codatbanking/models/operations/list_bank_transactions.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import transactions as shared_transactions
+from ...models.shared import transactions as shared_transactions
 from typing import Optional
 
 
@@ -27,10 +27,10 @@ class ListBankTransactionsRequest:
 class ListBankTransactionsResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     transactions: Optional[shared_transactions.Transactions] = dataclasses.field(default=None)
     r"""Success"""